requirements.txt
Django==5.1.6
djangorestframework==3.15.2
mysqlclient==2.2.5
requests==2.31.0
djangorestframework-simplejwt==5.5.0
python-dotenv==1.0.1
drf-spectacular==0.28.0
django-cors-headers==4.7.0
pytz==2025.2
cryptography==40.0.0
gunicorn==23.0.0